challenge-5/backend/api/requirements.txt
functions-framework
google-cloud-bigquery
pandas
google-genai
Flask
db-dtypes